\name{compute.runs.expectancy}
\alias{compute.runs.expectancy}

\title{
  Computes Runs Expectancies
}
\description{
  Given a raw Retrosheet play-by-play datafile stored as allYEAR.csv, will attach a header and compute the run expectancy values.
}
\usage{
  compute.runs.expectancy(season)
}
\arguments{
  \item{season}{
    numerical season value
  }
}

\value{
  Retrosheet data frame with new columns for computing run expectancies
}
\author{
  Jim Albert
}

