\name{compute.win.probs}
\alias{compute.win.probs}

\title{
  Computes Win Probabilities
}
\description{
  Given a Retrosheet play-by-play dataset with run values added, function will add new win probability values to dataset.
}
\usage{
  compute.win.probs(d)
}
\arguments{
  \item{d}{
    Retrosheet dataset with run values computed
  }
}

\value{
  Retrosheet data frame with new win probability columns
}
\author{
  Jim Albert
}

